add: class TextUtils and collect all utility functions there
